component(argument) `argument.hpp` contains a `constexpr` function for comparing C strings, the `_Arg` struct, and the two derived structs: `Flag` & `Option`.
`_Arg` contains most of the code that drives `Flag` and `Option` such as comparison operators, constructors, name return functions, etc.

`Flag` only has two unique functions, which are for setting the `_active` boolean to true or false.

`Option` is a little more complex, as it needs to keep track of its `_value` and `_is_option_mandatory` variables (which is why it handles its own constructors).